requirements.txt
# Core data science
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
scipy==1.11.4

# Deep learning
# TensorFlow is optional here because the autoencoder falls back to PCA
# when it is not available on the current Python/runtime combination.

# API backend
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0

# Visualization (for notebooks/reports)
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Data generation (for synthetic CERT-like data)
faker==22.0.0

# Utilities
joblib==1.3.2
python-dateutil==2.8.2
tqdm==4.66.1
colorlog==6.8.0
pyyaml==6.0.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
